Move test utils into the package.
