+++ packages/test-utils/src/index.ts
@@ -3,7 +3,5 @@
  * @license     https://opensource.org/licenses/MIT
  */
 
-export * from './types';
+export * from '@boost/core/test-utils';
 export * from './fixtures';
-export * from './mocks';
-export * from './stubs';
